implemented redux and using it to update admin/ tenant settings wip
